Consolidate Photos SQLite infrastructure into db.rs
